# src/generated/models/CreateRecordingDatePollRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateRecordingDatePollRequest = {
    title: string;
    candidateDates: Array<string>;
};